+++ src/utils/castArrayShape.js
@@ -1,0 +1,9 @@
+function castArrayShape(shouldBeArray, value) {
+  const isArray = Array.isArray(value);
+
+  return shouldBeArray ?
+    isArray ? value : [value] :
+    isArray ? value[0] : value;
+}
+
+module.exports = castArrayShape;