+++ src/utils/isNil.js
@@ -1,0 +1,5 @@
+function isNil(value) {
+  return typeof value === 'undefined' || value === null || (Array.isArray(value) && !value.length);
+}
+
+module.exports = isNil;