+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+# tbd
+
+**New features:**
+- Promise support for all resolvers
+
 # 0.1.0
 
 First release! :tada: